backoffice: Setup Prometheus with host and Apache metrics
Prometheus lets us easily track lots of useful metrics.  It comes with a
basic frontend for looking at the data, but Granfana is very commonly
layered on top for better dashboarding.  There is also a robust alerting
layer.  Neither are included here for now, but would be relatively quick
great future additions.

Prometheus' web configuration is intended to be proxied to by Apache,
which is separately managed in our seattleflu/backoffice-apache2 repo
(for now).