backoffice: Update apt cache before installing packages
This improves the clean-slate experience by ensuring the cache is filled
even if Ansible is run immediately after first boot when there's not yet
an apt cache.  Apt errors if there's no cache yet, so this removes a
necessary step which was previously manual, albeit at the cost of some
slight runtime increase.