Scaffold a convenient place to drop systemd units that should be installed
systemd will be automatically reloaded if any new files are added or
existing files changed.

Note that if you want to remove a unit file, a separate Ansible task is
necessary to clean it up.  This cannot be done automatically in a
reasonable and safe way with our current setup.