Set ACLs on the Switchboard data directory
We're enabling journal_mode=wal on the SQLite database for better
concurrency, but this means that Switchboard's web app now needs write
access to create and update the sidecar files used by SQLite for its WAL.

As the data updating cronjob runs as ubuntu and the data serving web app
runs as nobody, this access is easiest managed via ACLs instead of
standard owner/group perms.  (The alternative is creating a group to put
both users into and making files owned by that group.)

The expectation is that the rest of Switchboard's deployment will
eventually be managed via Ansible as well.  This is just a toehold.