Attach an "app" label to the metrics exported from uWSGI
We've previously relied upon the automatic "stats_uri" label to identify
the app.  This works, but makes for awkward metrics queries and ugly
graph labels in Grafana.  As a workaround, some graphs applied
additional transformation in Grafana to the stats_uri to extract the app
name.  This can all go away by exporting an explicit "app" label.  \o/