promtail: Listen on 9080 instead of a dynamic port
Dynamic ports make sense in a service mesh (e.g. k8s), but we don't use
a service mesh.  It would also make sense if we didn't want to scrape
metrics from promtail, but we will, so a static port makes Prometheus
config easier.  There's no official port for promtail, but 9080 is used
in several examples in the official docs, and it doesn't conflict with
any other Prometheus exporters.

We won't use promtail's gRPC interface but we can't disable it, so omit
it and let it default (9095).