promtail: Define a push destination in development
Otherwise promtail doesn't start and isn't fully testable in dev.  Since
we only have one Loki instance, this labels logs by "environment" so we
can filter out dev logs.

As a potential improvement for the future, we could register for a
free-tier Grafana Cloud account for dev/testing and push logs to that
second account instead of our production account.  This would be handy
for metrics as well.  Even nicer would be if we could have separate prod
vs. dev instances in the same Grafana Cloud account, which would show up
as separate data sources in Grafana.