Re-read systemd units from disk before any services are restarted/reloaded
After changes to the unit definition files (e.g. *.service) on disk,
systemd must re-read the files for the changes to take effect on the
service, even across starts/stops/restarts/reloads/etc of the service.
For example, new command line arguments in a service definition won't be
used until the definition files are re-read by systemd.  `systemctl
daemon-reload` is the manual equivalent of this.

Because of task order, our playbook notifies the systemd handler before
the "X is restarted/reloaded" handlers, but contrary to what I'd assumed
when introducing these, handlers run in the order they're defined, not
in the order they're notified.