Run our uWSGI apps as separate systemd services
Simplifies our setup by ditching the Debian/Ubuntu-specific multi-app
uWSGI framework and converging on systemd as our common service
supervisor.  One big upside is that now deploys only have to
restart/reload the updated app instead of all apps.

The individual uwsgi@*.service units for each app are still not enabled
or started or otherwise managed by Ansible yet, as doing so requires
bringing in a bit more of our deployment process first.  However, this
change helps ratchet towards that future.

Deployment of this change will not change the status of any services.
To switch from the single multi-app service to the individual app
services, with a brief period of downtime for each, run:

    systemctl stop uwsgi
    systemctl disable uwsgi

    for app in api-{production,testing} husky-musher; do
        systemctl enable uwsgi@$app
        systemctl start uwsgi@$app
    done