wip! manage uwsgi apps
To do this, I think we need to bring in a lot more of the deployment
process (which is our longer-term goal anyway).

I'm 50% sure the organization of tasks into tasks/source.yaml isn't the
right factoring for deploys.  I'm 90% sure that the bits in
tasks/uwsgi.yaml, tasks/python.yaml, and other files are a good
factoring, though.

Todos:

  - Merge in backoffice.git into this repo, with contents under a subdir
    (for now), while preserving git history. This will let files in there be
    managed/installed by Ansible instead of by running git pull on
    backoffice.

  - Port and adapt documented deploy steps into idempotent tasks in one
    or more Ansible playbooks.

  - Bring Apache and Shibboleth management under the wings of Ansible
    too. I would expect to merge in backoffice-apache2.git and
    backoffice-shibboleth.git into this repo too, along similar lines to
    (1).

There's a continuum here though, and each item above can be split into
separate, iterative steps that rachet towards the goal. That said,
upfront time aside, it might be easier to do things in one fell swoop to
reduce the overhead of coordinating changes.