Limit Apache2 Task Scope and Set Directory Permissions
This change limits Apache2 tasks to package installs, which will be
safer than what was there before. This change was tested locally and
installs packages that are already on prod. This is not an exhaustive
and/or inclusive list of the packages necessary to run our apache2
processes on backoffice.

This change also includes a workaround to install the apache2 directory
into the /etc production directory. We use agent forwarding to access
private directories, but you can't agent forward to the root user. Since
only the root user has write access to /etc, we copy a remote sourced
clone from the local runners /etc directory.

I really dislike this workaround but don't see another way to get the
repository into the correct location. Cloning to /opt and then copying
to /etc don't seem to work as anticipated and delete over 20 files (I
assume because of broken links).