Checkout all Backoffice Repos in `/opt` directory
Alters the backoffice repos task script to checkout just our repositories
stored in the `/opt` directory. This is a bit safer than also doing the
apache2 repo in `/etc`, which requires some special permissions as it is
owned by the root user. This is taken care of in a separate playbook.