Add all of our Systemd units to Systemd Task
We were already adding the uwsgi, promtail, and prometheus exporter with
our systemd configuration, but we were not adding the sfs-switchboard
service or linking the metabase or lab labels services from the backoffice
repo. This change adds tasks to link up these unit files and adds the
sfs-switchboard service file to the configuration (don't worry, it was copy
pasted from prod so we won't have a repeat uwsgi issue!).