Add Descriptive Tags to Run Related Tasks Together and Finalize Documentation
This commit finalizes documentation for changes made in previous commits. It
also adds tags to related tasks so that they may be run easily together as
part of a configuration bundle. These tags are as follows:

- `init`: each task is tagged with init, which allows you to run the
whole playbook easily upon initialization of an environment.
- `logging`: This tag is applied to tasks that touch configurations
to our monitoring services like promtail, prometheus, and logrotate.
- `deploy`: This tag is applied to tasks that fulfill duties within
our deployment pipeline. Running the playbook with tag deploy should
be able to pull in all regular changes to our code base.
- `dependencies`: This tag is applied to tasks that download system packages
or other dependencies from apt.
- `<task-name>`: Each task also has a tag unique to the task, allowing
minimal configuration changes to be deployed safely with high confidence.

We also add notify statements to each task import that might require a handler
restart or reload to ensure that they are run with any task that needs them.