Add Husky Musher to Ansible Playbook
This change adds our husky musher configuration to the Ansible playbook.
It sets up permissions on the musher directory, checks out the latest
version of the git repository, syncs the pipenv, sets up the log
directory with the correct permissions, creates a group for log file
ownership, and adds the necessary users to that user group.

musher fixup