Limit Postgres Tasks to Setup of Postgres Packages
This change eliminates the server building components of the original
postgres task list. This makes it safer to run on prod, since it should
only install the development headers and minimal packages required for
our Python interactions with a Postgres instance. With that said, like
with my recommendations for Python, I would only run this task list on
a fresh box so as not to assume the risk of messing with the Postgres
configuration.