Add a Resgistrations Task that Defines Necessary Variables
Since some of these tasks rely on variables set by others, but those
other tasks may not have been run due to the inclusion of task tags,
this change includes a registration step that always runs. The job
of this step is to register variables shared between tasks, ensuring
that all tasks can run independently of each other if desired.