Alter Tasks to Allow Run in Check and Diff Modes
Adds a few ansible built in options that allow the playbooks to be run in
check and diff mode. These modes in conjunction essentially allow you to run
playbooks as a dry run, while also outputting the changes that would have
been made on a normal invocation. Use `--diff --check` in conjunction.